+++ src/ra/netdlg.h
@@ -11,8 +11,7 @@ bool Remote_Connect();
 void Destroy_Connection(int id, int error);
 bool Process_Global_Packet(GlobalPacketType* packet, IPXAddressClass* address);
 uint32_t Compute_Name_CRC(char* name);
-void Net_Reconnect_Dialog(int reconn, int fresh, int oldest_index,
-                          unsigned long timeval);
+void Net_Reconnect_Dialog(int reconn, int fresh, int oldest_index, int timeval);
 
 extern const char* const* EngMisStr;
