+++ src/td/nullconn.h
@@ -100,8 +100,9 @@ class NullModemConnClass : public NonSequencedConnClass {
   /*.....................................................................
   Utility routines.
   .....................................................................*/
-  unsigned long Actual_Max_Packet() {
-    return MaxPacketLen + sizeof(SerialHeaderType) + sizeof(int) + sizeof(char);
+  int Actual_Max_Packet() {
+    return MaxPacketLen + static_cast<int>(sizeof(SerialHeaderType) +
+                                           sizeof(int) + sizeof(char));
   }
 
   /*.....................................................................
